Created a new function if RMlicelUSP.cpp to compare two profiles and verify if they can be added, averaged or merged.
The netcdf library now has a function to add a new profile into an existing netcdf file.
However, a copy of a RMDataFile variable written to that netcdf file is necessary in order to
check if they can be added togheter. In the future we could think of reading info from netcdf
and comparing it to the variable to be written.