Better solution for reaching the end of line. After doing a fscanf() for the exact format of whatever we need to read from a particular line, we to a
   fscanf(fp,"%*[^\n]"); fscanf(fp,"%*c");

which will read and discard everything until the next line feed (\n). The
second fscanf discards the line feed itself, so that the file-pointer is
actually pointing to the start of the next line.

It was necessary to add such double fscanf() after we read all channel lines
because there is an empty line before the raw data.